+++ django_mapengine/legend.py
@@ -3,7 +3,7 @@
 from dataclasses import dataclass
 from typing import Optional
 
-from . import layers, utils
+from . import utils
 
 
 @dataclass
@@ -12,13 +12,8 @@ class LegendLayer:
 
     name: str
     description: str
-    layer: layers.ModelLayer = None
-    layer_id: str = None
+    layer_id: str
     color: Optional[str] = None
-
-    def __post_init__(self):
-        if self.layer is None and self.layer_id is None:
-            raise ValueError("You must either set layer or layer_id.")
 
     def get_color(self) -> str:
         """
@@ -31,7 +26,7 @@ class LegendLayer:
         """
         if self.color:
             return self.color
-        return utils.get_color(self.get_layer_id())
+        return utils.get_color(self.layer_id)
 
     @property
     def style(self) -> dict:
@@ -43,33 +38,4 @@ class LegendLayer:
         dict
             layer style
         """
-        return utils.get_layer_style(self.get_layer_id())
-
-    def get_layer_id(self) -> str:
-        """
-        Return layer id either from layer or from user input
-
-        Returns
-        -------
-        str
-            Layer ID used in maplibre
-        """
-        if self.layer_id:
-            return self.layer_id
-        return self.layer.id
-
-    @property
-    def model(self):
-        """
-        Return related model
-
-        Either from stored layer or by looking up layer id in registered layers
-
-        Returns
-        -------
-        models.Model
-            Model related to current legend layer
-        """
-        if self.layer:
-            return self.layer.model
-        return layers.get_layer_by_id(self.get_layer_id()).model
+        return utils.get_layer_style(self.layer_id)